适配Linux系统
file-manager.py增加了判断请求是否为POST方法，没有这个的话浏览器直接访问cgi-bin/file-manager.py的时候程序会卡死，一直试图从标准输入中读取

+++ cgi-bin/file-manager.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 import cgitb
 import codecs
 import os
@@ -11,8 +13,8 @@ if not os.path.lexists(upload_dir):
     os.mkdir(upload_dir)
 # sys.excepthook = lambda exception_type, value, tb: print('<p>', html.escape(str(exception_type)), '<br>', value, '<br>', html.escape(str(tb)), '</p></body></html>')
 sys.stdout = codecs.getwriter('utf-8')(sys.stdout.buffer)
-print('Content-type: text/html; charset=utf-8')
-print()  # 空行，告诉服务器结束头部
+print('Content-type: text/html; charset=utf-8', end='\r\n')
+print(end='\r\n')  # 空行，告诉服务器结束头部
 print('<!DOCTYPE html>')
 print('<html lang="zh">')
 print('<head>')
@@ -24,17 +26,18 @@ print('<h2>文件管理</h2>')
 query = ''
 deleted = []
 non_existent = []
-try:
-    query = parse_qs(input())
-    for filename in query['filename']:
-        basename = os.path.basename(filename)
-        if os.path.exists(upload_dir + basename):
-            os.remove(upload_dir + basename)
-            deleted.append(basename)
-        else:
-            non_existent.append(basename)
-except EOFError:
-    pass  # 没有POST消息体，即没有要删除的文件
+if os.environ.get('REQUEST_METHOD') == 'POST':
+    try:
+        query = parse_qs(input())
+        for filename in query['filename']:
+            basename = os.path.basename(filename)
+            if os.path.exists(upload_dir + basename):
+                os.remove(upload_dir + basename)
+                deleted.append(basename)
+            else:
+                non_existent.append(basename)
+    except EOFError:
+        pass  # 没有POST消息体，即没有要删除的文件
 print('<form method="post">')
 print('<table border="1"><thead><tr><td></td><td><b>文件名</b></td><td><b>大小</b></td><td><b>上传时间</b></td></tr></thead><tbody>')
 with os.scandir(upload_dir) as file_list:
